requirements.txt
# Core framework
kivy==2.3.0
kivymd==1.1.1

# PDF generation
fpdf2==2.7.9

# Pillow (image support for Kivy/PDF)
Pillow==10.2.0

# Testing
pytest==8.1.1
pytest-cov==5.0.0
